requirements.txt
numpy 
pandas
scikit-learn
pymongo
certifi
python-dotenv
dill

-e .